requirements.txt
Flask==2.2.5
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Werkzeug==2.2.3
gunicorn==21.2.0
SQLAlchemy==1.4.51 